Very Basic Example Dashboard exported to JSON.
Someone on twitter asked for an example of the dashboard configs.  Honestly, I just use the GUI to create mine and don't do anything too fancy.  For a good walkthrough on how to create dashboards, I used the post by Chris Wahl [Building a dashboard with grafana and powercli](http://wahlnetwork.com/2015/04/29/building-a-dashboard-with-grafana-influxdb-and-powercli/).